✦ refactor(reading-tracker): modularize session recording view
Extract inline sheets, cards, and timer logic into dedicated component
and section widgets to improve maintainability and readability.
