+++ lib/features/reading_tracker/view/session_helpers.dart
@@ -1,0 +1,7 @@
+String formatSessionDuration(int totalSeconds) {
+  final duration = Duration(seconds: totalSeconds);
+  String twoDigits(int n) => n.toString().padLeft(2, '0');
+  final twoDigitMinutes = twoDigits(duration.inMinutes.remainder(60));
+  final twoDigitSeconds = twoDigits(duration.inSeconds.remainder(60));
+  return '${twoDigits(duration.inHours)}:$twoDigitMinutes:$twoDigitSeconds';
+}
